Fix Google Analytics adding during build procedure. Close #304

+++ dist.sh
@@ -1,9 +1,11 @@
 #!/usr/bin/env bash
+set -euo pipefail
 # Script for building tar.gz archive for deployment
 # author Oleg.Shpynov os@jetbrains.com
 
 VERSION=0.20
 BUILD=development
+GA=""
 
 for ARGUMENT in "$@"; do
     KEY=$(echo "${ARGUMENT}" | cut -f1 -d=)
@@ -25,15 +27,18 @@ PTV=pubtrends-${VERSION_BUILD}
 echo "Copy sources"
 mkdir -p "${PTV}"
 cp environment.yml "${PTV}"/
-cp Dockerfile "${PTV}"/
+cp resources/docker/main/Dockerfile "${PTV}"/
 cp config.properties "${PTV}"/
 cp docker-compose.yml "${PTV}"/
 cp gunicorn.sh "${PTV}"/
-cp -r pysrc "${PTV}"/
-cp -r pubtrends-review "${PTV}"/
+cp README.md ${PTV}/
+cp LICENSE.txt ${PTV}/
+rsync -aiz pysrc "${PTV}"/ --exclude=**/.git* --exclude=**/__pycache__*
+rsync -aiz pubtrends-review "${PTV}"/ --exclude=**/.git* --exclude=**/__pycache__*
 
 echo "Update config VERSION"
-sed -E "s/VERSION[^\n]*/VERSION = '${FULL_VERSION}'/g" -i "${PTV}"/pysrc/version.py
+# Don't use sed -i for BSD compatibility
+sed -E "s/VERSION[^\n]*/VERSION = '${FULL_VERSION}'/g" pysrc/version.py > "${PTV}"/pysrc/version.py
 
 echo "Setup GA"
 if [[ ! -z "${GA}" ]]; then
@@ -45,25 +50,24 @@ if [[ ! -z "${GA}" ]]; then
   gtag('js', new Date());\
   gtag('config', '$GA');\
 </script>"
-
+  echo "Google Analytics script: $GA_SCRIPT"
   # Ignore init.html file and admin/ files
-  for F in $(find "${PTV}"/pysrc -name "*.html" | grep -v 'init.html' | grep - v 'admin/'); do
+  for F in $(find "${PTV}"/pysrc -name "*.html" | grep -v init.html | grep -v admin/); do
     echo "Save version to $F"
-    sed "s#{{ version }}#${FULL_VERSION}#" -i $F
+    sed "s#{{ version }}#${FULL_VERSION}#" $F > $F.new
+    mv $F.new $F
     if [[ -z "$(echo $F | grep 'admin')" ]]; then
       echo "Adding Google Analytics tracker to $F"
-      echo "$GA_SCRIPT"
-      sed "s#<head>#<head>${GA_SCRIPT}#" -i $F
+      sed "s#<head>#<head>${GA_SCRIPT}#" $F > $F.new
+      mv $F.new $F
     fi
   done
 fi
 
-echo "Create folder for logs"
-mkdir "${PTV}"/logs
-chmod a+rwx "${PTV}"/logs
-
 echo "Create distributive dist/${PTV}.tar.gz"
-rm -r dist
+if [[ -d dist ]]; then
+  rm -r dist
+fi
 mkdir -p dist
 tar -zcf "dist/${PTV}.tar.gz" "${PTV}"
 
@@ -74,3 +78,5 @@ echo "Move jar to dist if exists"
 if [[ -f build/libs/pubtrends-dev.jar ]]; then
   mv build/libs/pubtrends-dev.jar "dist/${PTV}.jar"
 fi
+
+echo "Done"
